# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Turnpike Analyst Enterprise Platform API",
    "description": "Production-ready RESTful API backend for Turnpike Analyst covering ECM migrations (JAMES WEBB Server), AI Agent tools (Agent P8), training courses, consultations, CMS/blogging, lead capture, and role-based administration.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/training/courses": {
      "get": {
        "tags": [
          "Training Courses & Instructor Platform"
        ],
        "summary": "List Training Courses",
        "operationId": "list_training_courses_api_v1_training_courses_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CourseCategory"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by training category",
              "title": "Category"
            },
            "description": "Filter by training category"
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CourseLevel"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by course difficulty level",
              "title": "Level"
            },
            "description": "Filter by course difficulty level"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseResponse"
                  },
                  "title": "Response List Training Courses Api V1 Training Courses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/training/courses/{slug}": {
      "get": {
        "tags": [
          "Training Courses & Instructor Platform"
        ],
        "summary": "Get Course Curriculum Details",
        "operationId": "get_course_curriculum_details_api_v1_training_courses__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/training/become-instructor": {
      "post": {
        "tags": [
          "Training Courses & Instructor Platform"
        ],
        "summary": "Apply To Become Instructor",
        "operationId": "apply_to_become_instructor_api_v1_training_become_instructor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstructorApplicationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstructorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/training/instructor-applications": {
      "get": {
        "tags": [
          "Training Courses & Instructor Platform"
        ],
        "summary": "Review Instructor Applications",
        "operationId": "review_instructor_applications_api_v1_training_instructor_applications_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApplicationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by review status",
              "title": "Status"
            },
            "description": "Filter by review status"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InstructorApplicationResponse"
                  },
                  "title": "Response Review Instructor Applications Api V1 Training Instructor Applications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplicationStatus": {
        "type": "string",
        "enum": [
          "SUBMITTED",
          "UNDER_REVIEW",
          "ACCEPTED",
          "REJECTED"
        ],
        "title": "ApplicationStatus"
      },
...
      "CourseCategory": {
        "type": "string",
        "enum": [
          "ECM_TRAINING",
          "OCP_TRAINING",
          "CLOUD_DEVOPS",
          "AI_ML_TRAINING"
        ],
        "title": "CourseCategory"
      },
      "CourseLevel": {
        "type": "string",
        "enum": [
          "BEGINNER",
          "INTERMEDIATE",
          "ADVANCED"
        ],
        "title": "CourseLevel"
      },
      "CourseResponse": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 150,
            "minLength": 3,
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "maxLength": 150,
            "minLength": 3,
            "title": "Slug"
          },
          "category": {
            "$ref": "#/components/schemas/CourseCategory"
          },
          "duration_hours": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Duration Hours"
          },
          "level": {
            "$ref": "#/components/schemas/CourseLevel"
          },
          "curriculum": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Curriculum"
          },
          "instructor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructor Id"
          },
          "price": {
            "type": "number",
            "minimum": 0.0,
            "title": "Price"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published",
            "default": true
          },
          "_id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "slug",
          "category",
          "duration_hours",
          "level",
          "price"
        ],
        "title": "CourseResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InstructorApplicationCreate": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "expertise_areas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expertise Areas"
          },
          "resume_url": {
            "type": "string",
            "title": "Resume Url"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "email",
          "phone",
          "resume_url"
        ],
        "title": "InstructorApplicationCreate"
      },
      "InstructorApplicationResponse": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "expertise_areas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expertise Areas"
          },
          "resume_url": {
            "type": "string",
            "title": "Resume Url"
          },
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus",
            "default": "SUBMITTED"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "email",
          "phone",
          "resume_url"
        ],
        "title": "InstructorApplicationResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}